External_Apps/sustainablebrands/requirements.txt
altair
numpy
pandas
pydeck
streamlit
google.generativeai
plotly
